Moved some additional scanner classes
