+++ migration-cli/pom.xml
@@ -51,7 +51,7 @@
             <artifactId>quarkus-picocli</artifactId>
         </dependency>
 
-        <!-- Openrewrite -->
+        <!-- TODO: To be removed when PR 145 will be merged  -->
         <dependency>
             <groupId>org.openrewrite</groupId>
             <artifactId>rewrite-java</artifactId>
@@ -81,16 +81,6 @@
             <groupId>org.eclipse.lsp4j</groupId>
             <artifactId>org.eclipse.lsp4j</artifactId>
         </dependency>
-        <dependency>
-            <groupId>org.eclipse.lsp4j</groupId>
-            <artifactId>org.eclipse.lsp4j.jsonrpc</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>io.quarkus</groupId>
-            <artifactId>quarkus-junit5-mockito</artifactId>
-            <scope>test</scope>
-        </dependency>
-
     </dependencies>
 
     <build>
