+++ pom.xml
@@ -28,6 +28,7 @@
         <gson.version>2.13.2</gson.version>
         <ascii-table.version>1.8.0</ascii-table.version>
         <jackson.version>2.18.0</jackson.version>
+        <opencsv.version>5.12.0</opencsv.version>
 
         <compiler.plugin.version>3.14.1</compiler.plugin.version>
         <surefire.plugin.version>3.5.4</surefire.plugin.version>
@@ -112,7 +113,7 @@
             <dependency>
                 <groupId>com.opencsv</groupId>
                 <artifactId>opencsv</artifactId>
-                <version>5.12.0</version>
+                <version>${opencsv.version}</version>
             </dependency>
 
             <dependency>
