+++ tests/src/test/java/dev/snowdrop/analyze/properties/PropertiesFileSearchUsingScannerTest.java
@@ -1,11 +1,11 @@
-package dev.snowdrop.scanner.file;
+package dev.snowdrop.analyze.properties;
 
 import dev.snowdrop.analyze.BaseRulesTest;
 import dev.snowdrop.analyze.Config;
 import dev.snowdrop.analyze.model.Match;
 import dev.snowdrop.analyze.model.Rule;
-import dev.snowdrop.analyze.services.CodeScannerService;
-import dev.snowdrop.analyze.services.ScanCommandExecutor;
+import dev.snowdrop.service.scanner.CodeScannerService;
+import dev.snowdrop.service.scanner.ScanCommandExecutor;
 import dev.snowdrop.service.scanner.file.FileSearch;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.io.TempDir;
