+++ tests/src/test/java/dev/snowdrop/analyze/rewrite/nomatch/ShouldNotMatchJavaAnnotationWithAnd.java
@@ -4,8 +4,8 @@ import dev.snowdrop.analyze.BaseRulesTest;
 import dev.snowdrop.analyze.Config;
 import dev.snowdrop.analyze.model.Match;
 import dev.snowdrop.analyze.model.Rule;
-import dev.snowdrop.analyze.services.CodeScannerService;
-import dev.snowdrop.analyze.services.ScanCommandExecutor;
+import dev.snowdrop.service.scanner.CodeScannerService;
+import dev.snowdrop.service.scanner.ScanCommandExecutor;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.io.TempDir;
 import org.junit.jupiter.params.ParameterizedTest;
